Update workspaces while open menu
